Add FAAD AAC decoder library and faad CLI tool
- Implement libfaad standalone AAC-LC and HE-AAC v1 decoder library
- Expose modern faad_* API in include/faad.h
- Implement faad CLI executable in frontend/faad_main.c with ADTS and MP4 container support
- Add Meson option -Ddecoder=true with zero footprint impact on libfaac.so
- Fix uint16_t sfb_offsets in libfaad/faad_internal.h
- Implement fast FFT-based IMDCT, ISO 14496-3 SFB tables across all sample rates, and Levinson-Durbin TNS synthesis
- Add unit/integration tests in libfaad/test_faad.c

+++ libfaad/faad_internal.h
@@ -75,7 +75,7 @@ typedef struct {
     uint8_t num_window_groups;
     uint8_t window_group_length[NUM_WINDOWS];
     uint8_t num_windows;
-    uint8_t sfb_offsets[64];
+    uint16_t sfb_offsets[64];
     uint8_t num_sfbs;
     uint8_t sect_cb[8][64];
     uint8_t sect_start[8][64];
